Port PaperMC/Paper#4101
This is a reimplementation of AIkar's villager optimizations from 1.15.2 that have been updated to work with 1.16.3. This makes a large difference to villager performance.

This pull request will likely be merged into Paper at a future date (once all requested changes have been made), at which point it will be pulled from this project.